Moved PDFs to neutrinos_irfs + optimize PDF calls
