+++ src/momenta/io/neutrinos.py
@@ -25,16 +25,14 @@ import healpy as hp
 import numpy as np
 import scipy.integrate
 import yaml
-from scipy.integrate import quad, trapezoid
-from scipy.interpolate import interp1d, RegularGridInterpolator
 from scipy.linalg import block_diag
 from scipy.stats import gamma, truncnorm
 
 import astropy.coordinates
 import astropy.time
-from astropy.units import deg, rad
+from astropy.units import rad
 
-import momenta.stats.pdfs as pdf
+import momenta.io.neutrinos_irfs as irfs
 
 
 warnings.filterwarnings("ignore", category=scipy.integrate.IntegrationWarning)
@@ -61,145 +59,6 @@ def infer_uncertainties(input_array: float | np.ndarray, nsamples: int, correlat
     if input_array.shape == (nsamples, nsamples):
         return input_array
     raise RuntimeError("The size of uncertainty_acceptance does not match with the number of samples")
-
-
-class EffectiveAreaBase:
-    """Class to handle detector effective area for a given sample and neutrino flavour.
-    This default class handles only energy-dependent effective area."""
-
-    def __init__(self):
-        self._acceptances = {}
-
-    def evaluate(self, energy: float | np.ndarray, ipix: int, nside: int):
-        """Evaluate the effective area for a given energy, pixel index, and skymap resolution."""
-        return 0
-
-    def _compute_acceptance(self, fluxcomponent, ipix: int, nside: int):
-        """Compute the acceptance integrating the effective area x flux in log scale between emin and emax.
-        Only used internally by `compute_acceptance_map`, may be overriden in a inheriting class."""
-        def func(x: float):
-            return fluxcomponent.evaluate(np.exp(x)) * self.evaluate(np.exp(x), ipix, nside) * np.exp(x)
-
-        return quad(func, np.log(fluxcomponent.emin), np.log(fluxcomponent.emax), limit=500)[0]
-
-    def compute_acceptance_map(self, fluxcomponent, nside: int):
-        """Compute the acceptance map for a given flux component, iterating over all pixels."""
-        return np.array([self._compute_acceptance(fluxcomponent, ipix, nside) for ipix in range(hp.nside2npix(nside))])
-
-    def compute_acceptance_maps(self, fluxcomponents, nside: int):
-        """Compute the acceptance maps for a list of flux components, iterating over all components and pixels.
-        May be overriden by a smarter implementation for specific cases where computation can be optimized."""
-        return [self.compute_acceptance_map(c, nside) for c in fluxcomponents]
-
-    def get_acceptance_map(self, fluxcomponent, nside: int):
-        """Get the acceptance for a given flux component. 
-        If the component `store` attribute is 'exact', it is retrieved from the `acceptances` dictionary (added there if not yet available).
-        If the component `store` attribute is 'interpolate', the dictionary with the interpolation function (+inputs) is returned.
-        """
-        if fluxcomponent.store == "exact":
-            if (str(fluxcomponent), nside) not in self._acceptances:
-                self._acceptances[(str(fluxcomponent), nside)] = self.compute_acceptance_map(fluxcomponent, nside)
-            return self._acceptances[(str(fluxcomponent), nside)]
-        if fluxcomponent.store == "interpolate":
-            if (str(fluxcomponent), nside) not in self._acceptances:
-                accs = {str(c): a for c, a in zip(fluxcomponent.grid.flatten(), self.compute_acceptance_maps(fluxcomponent.grid.flatten(), nside))}
-                def f(_c):
-                    return accs[str(_c)]
-                accs = np.vectorize(f, signature="()->(n)")(fluxcomponent.grid)
-                grid = [*fluxcomponent.shapevar_grid, np.arange(hp.nside2npix(nside))]
-                self._acceptances[(str(fluxcomponent), nside)] = RegularGridInterpolator(grid, accs)
-            return self._acceptances[(str(fluxcomponent), nside)]
-        return self.compute_acceptance_map(fluxcomponent, nside)
-
-    def get_acceptance(self, fluxcomponent, ipix: int, nside: int):
-        """Get the acceptance"""
-        if fluxcomponent.store == "exact":
-            return self.get_acceptance_map(fluxcomponent, nside)[ipix]
-        if fluxcomponent.store == "interpolate":
-            return self.get_acceptance_map(fluxcomponent, nside)([*fluxcomponent.shapevar_values, ipix])
-        return self._compute_acceptance(fluxcomponent, ipix, nside)
-
-
-class EffectiveAreaAllSky(EffectiveAreaBase):
-
-    def __init__(self):
-        super().__init__()
-        self.func = None
-
-    def read_csv(self, csvfile: str):
-        x, y = np.loadtxt(csvfile, delimiter=",").T
-        self.func = interp1d(x, y, bounds_error=False, fill_value=0)
-
-    def evaluate(self, energy: float | np.ndarray, ipix: int, nside: int):
-        return self.func(energy)
-
-    def compute_acceptance_map(self, fluxcomponent, nside):
-        acc = self._compute_acceptance(fluxcomponent, 0, nside) * np.ones(hp.nside2npix(nside))
-        return acc
-
-
-class EffectiveAreaDeclinationDep(EffectiveAreaBase):
-
-    def __init__(self):
-        super().__init__()
-        self.mapping = {}
-
-    def evaluate(self, energy: float | np.ndarray, ipix: int, nside: int):
-        if nside not in self.mapping:
-            self.mapping[nside] = self.map_ipix_to_declination(nside)
-        return self.func(energy, self.mapping[nside][ipix])
-
-    def compute_acceptance_map(self, fluxcomponent, nside):
-        if nside not in self.mapping:
-            self.mapping[nside] = self.map_ipix_to_declination(nside)
-        acc = np.zeros(hp.nside2npix(nside))
-        for dec, ipix in zip(*np.unique(self.mapping[nside], return_index=True)):
-            acc[self.mapping[nside] == dec] = self._compute_acceptance(fluxcomponent, ipix, nside)
-        return acc
-
-    def map_ipix_to_declination(self, nside):
-        ipix = np.arange(hp.nside2npix(nside))
-        _, dec = hp.pix2ang(nside, ipix, lonlat=True)
-        return dec
-
-
-class EffectiveAreaAltitudeDep(EffectiveAreaBase):
-
-    def __init__(self):
-        super().__init__()
-        self.func = None
-        self.mapping = {}
-
-    def read(self):
-        bins_logenergy = ...  # shape (M,)
-        bins_altitude = ...  # shape (N,)
-        aeff = ...  # shape (M,N)
-        self.func = RegularGridInterpolator((bins_logenergy, bins_altitude), aeff, bounds_error=False, fill_value=0)
-
-    def set_location(self, time, lat_deg, lon_deg):
-        self.obstime = time
-        self.location = astropy.coordinates.EarthLocation(lat=lat_deg * deg, lon=lon_deg * deg)
-        self.mapping = {}
-
-    def evaluate(self, energy: float | np.ndarray, ipix: int, nside: int):
-        if nside not in self.mapping:
-            self.mapping[nside] = self.map_ipix_to_altitude(nside)
-        return self.func((np.log10(energy), self.mapping[nside][ipix]))
-
-    def compute_acceptance_map(self, fluxcomponent, nside):
-        if nside not in self.mapping:
-            self.mapping[nside] = self.map_ipix_to_altitude(nside)
-        acc = np.zeros(hp.nside2npix(nside))
-        for alt, ipix in zip(*np.unique(self.mapping[nside], return_index=True)):
-            acc[self.mapping[nside] == alt] = self._compute_acceptance(fluxcomponent, ipix, nside)
-        return acc
-
-    def map_ipix_to_altitude(self, nside):
-        ipix = np.arange(hp.nside2npix(nside))
-        ra, dec = hp.pix2ang(nside, ipix, lonlat=True)
-        coords_eq = astropy.coordinates.SkyCoord(ra=ra * deg, dec=dec * deg, frame="icrs")
-        coords_loc = coords_eq.transform_to(astropy.coordinates.AltAz(obstime=self.obstime, location=self.location))
-        return coords_loc.alt.deg
 
 
 class Background(abc.ABC):
@@ -345,12 +204,12 @@ class NuSample:
 
     def set_pdfs(
         self,
-        sig_ang: pdf.AngularSignal | None = None,
-        sig_ene: pdf.EnergySignal | None = None,
-        sig_time: pdf.TimeSignal | None = None,
-        bkg_ang: pdf.AngularBackground | None = None,
-        bkg_ene: pdf.EnergyBackground | None = None,
-        bkg_time: pdf.TimeBackground | None = None,
+        sig_ang: irfs.AngularSignal | None = None,
+        sig_ene: irfs.EnergySignal | None = None,
+        sig_time: irfs.TimeSignal | None = None,
+        bkg_ang: irfs.AngularBackground | None = None,
+        bkg_ene: irfs.EnergyBackground | None = None,
+        bkg_time: irfs.TimeBackground | None = None,
     ):
         if (sig_ang is None) ^ (bkg_ang is None):
             raise RuntimeError("One of the angular PDFs is missing!")
@@ -448,7 +307,7 @@ class NuDetector(NuDetectorBase):
         for i, smp in enumerate(self.samples):
             smp.set_observations(nobserved[i], background[i])
 
-    def set_effective_areas(self, aeffs: list[EffectiveAreaBase]):
+    def set_effective_areas(self, aeffs: list[irfs.EffectiveAreaBase]):
         for i, smp in enumerate(self.samples):
             smp.set_effective_area(aeffs[i])
